+++ ios/RNInstabug/RCTConvert+InstabugEnums.m
@@ -109,4 +109,11 @@ RCT_ENUM_CONVERTER(
   integerValue
 );
 
+RCT_ENUM_CONVERTER(
+  IBGActionType,
+  ArgsRegistry.userConsentActionTypes,
+  IBGActionTypeNoChat,
+  integerValue
+);
+
 @end